SASS-11074: Add this FormProvider trait to support further generalisation. Some minor refactoring
